requirements.txt
twitchio==2.9.1
python-dotenv==1.0.0
requests==2.31.0